creates logging_utilities and moves functions outside of main app file
